refactor(llm): decompose GroqLlamaAdapter to match its siblings
The OpenAI adapter is 510 lines plus five submodules; Gemini is 409 plus
four. Groq-Llama was 1,046 with everything inline — the same responsibility
with none of the decomposition.

Three cohesive clusters move out: message assembly, response normalization,
and context budgeting, plus the request/response types (mirroring
openai/types.ts). All were pure apart from one logger reference, which is now
a parameter, so the extraction is mechanical rather than a rewrite. 1,046 ->
713 with four submodules.

Transport stays put. _executeRequest (272 lines), streamComplete (202) and
complete (103) are the real bulk, and restructuring the hot path of the
provider that serves both active-loop surfaces — guarded by three tests — is
not something to do as a layout cleanup.

contextBudget gains the eight tests it never had, which is the actual reason
to extract it rather than the line count. Getting calculateMaxTokens wrong
either truncates JSON responses or lets a Llama 3 generation loop run away,
and neither shows up as a type error.
